% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch_jupyter}
\alias{launch_jupyter}
\title{Launch jupyter}
\usage{
launch_jupyter(path = ".")
}
\arguments{
\item{path}{Path to desired default directory for Jupyter environment.}
}
\value{
Opens a new terminal and executes a jupyter command, which should,
in turn, open the system's default browser to the launched jupyter server.
}
\description{
Opens terminal and launches a Juypter environment in the system's
default browser.
}
\section{Requirements}{

This function assumes the installation of
\href{https://www.anaconda.com/distribution/}{Anaconda} (a data science
distribution/platform for R and python), which comes bundled with
\href{https://www.python.org}{Python} and
\href{https://jupyter.org/about}{Jupyter}.

For instructions on installing Anaconda see
\url{http://docs.anaconda.com/anaconda/install/}.
}

